// types.ts
export interface User {
  id: string;
  nome: string;
  cognome: string;
  genere: string;
  eta: number;
  codiceFiscale?: string;
}

export interface Book {
  id: string;
  titolo: string;
  autore: string;
  anno: number;
  genere: string;
  isbn: string;
}

export interface Loan {
  id: string;
  userId: string;
  bookId: string;
  dataPrestito: string; // ISO 8601 format
  dataRestituzione: string | null; // ISO 8601 format or null
}

export type View = 'dashboard' | 'books' | 'users' | 'loans';

export interface LibraryData {
    users: User[];
    books: Book[];
    loans: Loan[];
    addUser: (user: Omit<User, 'id'>) => void;
    updateUser: (user: User) => void;
    deleteUser: (userId: string) => void;
    addBook: (book: Omit<Book, 'id'>) => void;
    updateBook: (book: Book) => void;
    deleteBook: (bookId: string) => void;
    // Fix: Changed 'addLoan' parameter type to not require 'dataPrestito' as it's generated by the function.
    addLoan: (loan: Omit<Loan, 'id' | 'dataRestituzione' | 'dataPrestito'>) => void;
    returnLoan: (loanId: string) => void;
    deleteLoan: (loanId: string) => void;
}